+++ examples/c/.gitignore
